# docs/openapi.json
openapi: 3.0.3
info:
  title: Kassenbuch App v2 API
  version: 2.0.0
  description: REST API für die Kassenbuch App v2 – Verwaltung von Klassenkassen für
    Freizeitpädagoginnen. Django + Next.js Headless Architecture.
paths:
...
  /api/v1/groups/groups/{id}/members/:
    get:
      operationId: groups_groups_members_retrieve
      description: List all members of a group.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Ein eindeutiger Ganzzahl-Wert, der Gruppe identifiziert.
        required: true
      tags:
      - groups
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupList'
          description: ''
...
  /api/v1/groups/groups/{id}/members/add/:
    post:
      operationId: groups_groups_members_add_create
      description: Add a member to a group.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Ein eindeutiger Ganzzahl-Wert, der Gruppe identifiziert.
        required: true
      tags:
      - groups
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupListRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GroupListRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GroupListRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupList'
          description: ''
  /api/v1/groups/groups/{id}/students/:
    get:
      operationId: groups_groups_students_retrieve
      description: List all students of a group.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Ein eindeutiger Ganzzahl-Wert, der Gruppe identifiziert.
        required: true
      tags:
      - groups
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupList'
          description: ''
  /api/v1/groups/members/:
    get:
      operationId: groups_members_list
      description: |-
        CRUD for group members.

        - Educators: read-only
        - LocationManager+: full CRUD
      parameters:
      - name: ordering
        required: false
        in: query
        description: Feld, das zum Sortieren der Ergebnisse verwendet werden soll.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Eine Seitenzahl in der paginierten Ergebnismenge.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: Ein Suchbegriff.
        schema:
          type: string
      tags:
      - groups
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGroupMemberList'
          description: ''
    post:
      operationId: groups_members_create
      description: |-
        CRUD for group members.

        - Educators: read-only
        - LocationManager+: full CRUD
      tags:
      - groups
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupMemberCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GroupMemberCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GroupMemberCreateRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMemberCreate'
          description: ''
...
  /api/v1/groups/students/:
    get:
      operationId: groups_students_list
      description: |-
        CRUD for students.

        - Educators: read access to students in their groups
        - LocationManager+: full CRUD
      parameters:
      - in: query
        name: group_id
        schema:
          type: number
      - in: query
        name: is_active
        schema:
          type: boolean
      - name: ordering
        required: false
        in: query
        description: Feld, das zum Sortieren der Ergebnisse verwendet werden soll.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Eine Seitenzahl in der paginierten Ergebnismenge.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: Ein Suchbegriff.
        schema:
          type: string
      tags:
      - groups
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedStudentListList'
          description: ''
    post:
      operationId: groups_students_create
      description: |-
        CRUD for students.

        - Educators: read access to students in their groups
        - LocationManager+: full CRUD
      tags:
      - groups
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentCreate'
          description: ''
...
components:
  schemas:
...
    GroupList:
      type: object
      description: Serializer for listing groups (compact).
      properties:
        id:
          type: integer
          readOnly: true
        location:
          type: integer
          title: Standort
        location_name:
          type: string
          readOnly: true
        school_year:
          type: integer
          title: Schuljahr
        school_year_name:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
          title: Beschreibung
        leader:
          allOf:
          - $ref: '#/components/schemas/UserCompact'
          readOnly: true
        member_count:
          type: integer
          readOnly: true
        student_count:
          type: integer
          readOnly: true
        balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          title: Kontostand
          description: Aktueller Kontostand der Gruppenkasse in EUR
        currency:
          type: string
          title: Waehrung
          maxLength: 3
        is_active:
          type: boolean
          title: Aktiv
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: Erstellt am
      required:
      - balance
      - created_at
      - id
      - leader
      - location
      - location_name
      - member_count
      - name
      - school_year
      - school_year_name
      - student_count
    GroupListRequest:
      type: object
      description: Serializer for listing groups (compact).
      properties:
        location:
          type: integer
          title: Standort
        school_year:
          type: integer
          title: Schuljahr
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          title: Beschreibung
        currency:
          type: string
          minLength: 1
          title: Waehrung
          maxLength: 3
        is_active:
          type: boolean
          title: Aktiv
      required:
      - location
      - name
      - school_year
    GroupMember:
      type: object
      description: Serializer for group member display.
      properties:
        id:
          type: integer
          readOnly: true
        user_id:
          type: integer
          readOnly: true
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
        user_role:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/Role380Enum'
          title: Rolle in Gruppe
        is_active:
          type: boolean
          title: Aktiv
        joined_at:
          type: string
          format: date-time
          readOnly: true
          title: Beigetreten am
        left_at:
          type: string
          format: date-time
          nullable: true
          title: Ausgetreten am
      required:
      - first_name
      - id
      - joined_at
      - last_name
      - user_id
      - user_role
    GroupMemberCreate:
      type: object
      description: Serializer for adding a member to a group.
      properties:
        id:
          type: integer
          readOnly: true
        group:
          type: integer
          title: Gruppe
        user:
          type: integer
          title: Benutzer
        role:
          allOf:
          - $ref: '#/components/schemas/Role380Enum'
          title: Rolle in Gruppe
      required:
      - group
      - id
      - user
    GroupMemberCreateRequest:
      type: object
      description: Serializer for adding a member to a group.
      properties:
        group:
          type: integer
          title: Gruppe
        user:
          type: integer
          title: Benutzer
        role:
          allOf:
          - $ref: '#/components/schemas/Role380Enum'
          title: Rolle in Gruppe
      required:
      - group
      - user
...
    PaginatedGroupMemberList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
...
    PaginatedStudentListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/StudentList'
...
    Role380Enum:
      enum:
      - educator
      - assistant
      - substitute
      type: string
      description: |-
        * `educator` - Paedagogin
        * `assistant` - Assistenz
        * `substitute` - Vertretung
...
    StudentCreate:
      type: object
      description: Serializer for creating/updating students.
      properties:
        id:
          type: integer
          readOnly: true
        group:
          type: integer
          title: Gruppe
        first_name:
          type: string
          title: Vorname
          maxLength: 150
        last_name:
          type: string
          title: Nachname
          maxLength: 150
        date_of_birth:
          type: string
          format: date
          nullable: true
          title: Geburtsdatum
        email:
          type: string
          format: email
          title: E-Mail
          maxLength: 254
        phone:
          type: string
          title: Telefon
          maxLength: 20
        street:
          type: string
          title: Strasse
          maxLength: 255
        city:
          type: string
          title: Stadt
          maxLength: 100
        postal_code:
          type: string
          title: PLZ
          maxLength: 10
        is_active:
          type: boolean
          title: Aktiv
      required:
      - first_name
      - group
      - id
      - last_name
    StudentCreateRequest:
      type: object
      description: Serializer for creating/updating students.
      properties:
        group:
          type: integer
          title: Gruppe
        first_name:
          type: string
          minLength: 1
          title: Vorname
          maxLength: 150
        last_name:
          type: string
          minLength: 1
          title: Nachname
          maxLength: 150
        date_of_birth:
          type: string
          format: date
          nullable: true
          title: Geburtsdatum
        email:
          type: string
          format: email
          title: E-Mail
          maxLength: 254
        phone:
          type: string
          title: Telefon
          maxLength: 20
        street:
          type: string
          title: Strasse
          maxLength: 255
        city:
          type: string
          title: Stadt
          maxLength: 100
        postal_code:
          type: string
          title: PLZ
          maxLength: 10
        is_active:
          type: boolean
          title: Aktiv
      required:
      - first_name
      - group
      - last_name
...
    StudentList:
      type: object
      description: Serializer for listing students.
      properties:
        id:
          type: integer
          readOnly: true
        group:
          type: integer
          title: Gruppe
        group_name:
          type: string
          readOnly: true
        first_name:
          type: string
          title: Vorname
          maxLength: 150
        last_name:
          type: string
          title: Nachname
          maxLength: 150
        date_of_birth:
          type: string
          format: date
          nullable: true
          title: Geburtsdatum
        email:
          type: string
          format: email
          title: E-Mail
          maxLength: 254
        phone:
          type: string
          title: Telefon
          maxLength: 20
        is_active:
          type: boolean
          title: Aktiv
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: Erstellt am
      required:
      - created_at
      - first_name
      - group
      - group_name
      - id
      - last_name
...
    UserCompact:
      type: object
      description: Compact user representation for nested display.
      properties:
        id:
          type: integer
          readOnly: true
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
      required:
      - first_name
      - id
      - last_name
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...